audio mixer not somewhat working conversion code working but some issue with it distorting audio ffmpeg filter now includes audio on output, some issue exists that cause it to sometimes create invalid files. Also had lots of issues trying to connect to youtube

+++ libraries/medialib/medialib/audioConversion.h
@@ -91,9 +91,9 @@ size_t convertAudioBuffer(const AudioBuffer &audioBufferFrom, AudioBuffer &audio
 	break;
 	case AudioFormat::Float:
 	{
-		std::vector<uint8_t *> buffer;
+		std::vector<float *> buffer;
 
-		buffer.push_back((uint8_t *)audioBufferTo.getBuffer());
+		buffer.push_back((float *)audioBufferTo.getBuffer());
 		return convert<_FromType>(audioBufferFrom, 0, audioBufferFrom.getSamples(), buffer, audioBufferTo.getSize());
 	}
 	break;
